+++ themes-default/slim/src/components/app-header.vue
@@ -173,6 +173,17 @@ export default {
     mounted() {
         const { $el } = this;
 
+        // Auto close menus when clicking a RouterLink
+        $el.clickCloseMenus = event => {
+            const { target } = event;
+            if (target.matches('#main_nav a.router-link, #main_nav a.router-link *')) {
+                const dropdown = target.closest('.dropdown');
+                dropdown.querySelector('.dropdown-toggle').setAttribute('aria-expanded', false);
+                dropdown.querySelector('.dropdown-menu').style.display = 'none';
+            }
+        };
+        $el.addEventListener('click', $el.clickCloseMenus, { passive: true });
+
         // Hover Dropdown for Nav
         $($el).on({
             mouseenter(event) {
@@ -197,6 +208,22 @@ export default {
                     window.location.href = $target.attr('href');
                 }
             });
+        }
+    },
+    destroyed() {
+        // Revert `mounted()`
+        const { $el } = this;
+
+        // Auto close menus when clicking a RouterLink
+        $el.removeEventListener('click', $el.clickCloseMenus);
+
+        // Hover Dropdown for Nav
+        $($el).off('mouseenter mouseleave', 'ul.nav li.dropdown');
+
+        // @TODO Replace this with a real touchscreen check
+        // hack alert: if we don't have a touchscreen, and we are already hovering the mouse, then click should link instead of toggle
+        if ((navigator.maxTouchPoints || 0) < 2) {
+            $($el).off('click', '.dropdown-toggle');
         }
     },
     methods: {