+++ themes-default/slim/src/components/sub-menu.vue
@@ -1,67 +1,62 @@
-<script type="text/x-template" id="sub-menu-template">
-<div id="sub-menu-wrapper">
-    <div id="sub-menu-container" class="row shadow">
-        <div id="sub-menu" class="submenu-default hidden-print col-md-12">
-            <template v-for="menuItem in subMenu">
-                <app-link v-if="!menuItem.confirm" :key="menuItem.title" :href="menuItem.path" class="btn-medusa top-5 bottom-5">
+<template>
+    <div v-if="subMenu.length > 0" id="sub-menu-wrapper">
+        <div id="sub-menu-container" class="row shadow">
+            <div id="sub-menu" class="submenu-default hidden-print col-md-12">
+                <app-link
+                    v-for="menuItem in subMenu"
+                    :key="`sub-menu-${menuItem.title}`"
+                    :href="menuItem.path"
+                    class="btn-medusa top-5 bottom-5"
+                    @[clickEventCond(menuItem)].native.prevent="confirmDialog($event, menuItem.confirm)"
+                >
                     <span :class="['pull-left', menuItem.icon]"></span> {{ menuItem.title }}
                 </app-link>
-                <app-link v-else :key="menuItem.title" :href="menuItem.path" class="btn-medusa top-5 bottom-5" @click.native.prevent="confirmDialog($event, menuItem.class)">
-                    <span :class="['pull-left', menuItem.icon]"></span> {{ menuItem.title }}
-                </app-link>
-            </template>
 
-            <show-selector v-if="showSelectorVisible" :show-slug="curShowSlug" follow-selection></show-selector>
+                <show-selector v-if="showSelectorVisible" :show-slug="curShowSlug" follow-selection></show-selector>
+            </div>
         </div>
-    </div>
 
-    <!-- This fixes some padding issues on screens larger than 1281px -->
-    <div class="btn-group"></div>
-</div>
-</script>
-<%!
-    import json
-%>
+        <!-- This fixes some padding issues on screens larger than 1281px -->
+        <div class="btn-group"></div>
+    </div>
+</template>
 <script>
-const SubMenuComponent = {
+import { AppLink, ShowSelector } from './helpers';
+
+export default {
     name: 'sub-menu',
-    template: '#sub-menu-template',
-    data() {
-        return {
-            // Python conversions
-            // @TODO: Add the submenu definitions to VueRouter's routes object
-            rawSubMenu: ${json.dumps(submenu)}
-        };
+    components: {
+        AppLink,
+        ShowSelector
     },
     computed: {
         subMenu() {
-            return this.rawSubMenu.filter(item => item.requires === undefined || item.requires);
+            const { $route } = this;
+            let subMenu = $route.meta.subMenu || [];
+            if (typeof subMenu === 'function') {
+                subMenu = subMenu.call(this);
+            }
+            // Filters `requires = false` and reverses the array
+            const reducer = (arr, item) => (item.requires === undefined || item.requires) ? arr.concat(item) : arr;
+            return subMenu.reduceRight(reducer, []);
         },
         showSelectorVisible() {
-            const { pathname } = window.location;
-            return pathname.includes('/home/displayShow');
+            const { $route } = this;
+            return $route.name === 'show';
         },
         curShowSlug() {
-            if (!this.showSelectorVisible) {
-                return null;
-            }
-            const { params } = this;
-            const { indexername, seriesid } = params;
-            if (indexername && seriesid) {
+            const { $route, showSelectorVisible } = this;
+            const { indexername, seriesid } = $route.query;
+            if (showSelectorVisible && indexername && seriesid) {
                 return indexername + seriesid;
             }
             return '';
-        },
-        params() {
-            const { search } = window.location;
-            return search.slice(1).split('&').reduce((obj, pair) => {
-                const [key, value] = pair.split('=');
-                obj[key] = value;
-                return obj;
-            }, {});
         }
     },
     methods: {
+        clickEventCond(menuItem) {
+            return menuItem.confirm ? 'click' : null;
+        },
         confirmDialog(event, action) {
             const options = {
                 confirmButton: 'Yes',
@@ -77,12 +72,10 @@ const SubMenuComponent = {
             if (action === 'removeshow') {
                 const showName = document.querySelector('#showtitle').dataset.showname;
                 options.title = 'Remove Show';
-                <%text>
                 options.text = `Are you sure you want to remove <span class="footerhighlight">${showName}</span> from the database?<br><br>
                                 <input type="checkbox" id="deleteFiles"> <span class="red-text">Check to delete files as well. IRREVERSIBLE</span>`;
-                </%text>
                 options.confirm = $element => {
-                    window.location.href = $element[0].href + (document.getElementById('deleteFiles').checked ? '&full=1' : '');
+                    window.location.href = $element[0].href + (document.querySelector('#deleteFiles').checked ? '&full=1' : '');
                 };
             } else if (action === 'clearhistory') {
                 options.title = 'Clear History';
@@ -92,9 +85,9 @@ const SubMenuComponent = {
                 options.text = 'Are you sure you want to trim all download history older than 30 days?';
             } else if (action === 'submiterrors') {
                 options.title = 'Submit Errors';
-                options.text =  `Are you sure you want to submit these errors?<br><br>
-                                 <span class="red-text">Make sure Medusa is updated and trigger<br>
-                                 this error with debug enabled before submitting</span>`;
+                options.text = `Are you sure you want to submit these errors?<br><br>
+                                <span class="red-text">Make sure Medusa is updated and trigger<br>
+                                this error with debug enabled before submitting</span>`;
             } else {
                 return;
             }
@@ -103,8 +96,6 @@ const SubMenuComponent = {
         }
     }
 };
-
-window.components.push(SubMenuComponent);
 </script>
 <style scoped>
 /* Theme-specific styling adds the rest */