+++ themes-default/slim/src/utils.js
@@ -1,12 +1,12 @@
-const isDevelopment = process.env.NODE_ENV === 'development';
+export const isDevelopment = process.env.NODE_ENV === 'development';
 
 /**
  * Calculate the combined value of the selected qualities.
  * @param {number[]} allowedQualities - Array of allowed qualities.
- * @param {number[]} preferredQualities - Array of preferred qualities.
- * @returns {number} - An unsigned integer.
+ * @param {number[]} [preferredQualities=[]] - Array of preferred qualities.
+ * @returns {number} An unsigned integer.
  */
-const combineQualities = (allowedQualities, preferredQualities) => {
+export const combineQualities = (allowedQualities, preferredQualities = []) => {
     const reducer = (accumulator, currentValue) => accumulator | currentValue;
     const allowed = allowedQualities.reduce(reducer, 0);
     const preferred = preferredQualities.reduce(reducer, 0);
@@ -20,7 +20,7 @@ const combineQualities = (allowedQualities, preferredQualities) => {
  * @param {boolean} [useDecimal=false] - Use decimal instead of binary prefixes (e.g. kilo = 1000 instead of 1024)
  * @returns {string} The converted size.
  */
-const humanFileSize = (bytes, useDecimal = false) => {
+export const humanFileSize = (bytes, useDecimal = false) => {
     if (bytes === undefined) {
         return;
     }
@@ -39,12 +39,13 @@ const humanFileSize = (bytes, useDecimal = false) => {
     return `${bytes.toFixed(2)} ${units[u]}`;
 };
 
+
 /**
  * Map dateformat of pythons datetime.strftime() to that of javascripts dateFns.
  * @param {String} dateFormatString - pythons strftime format
  * @returns {String} mapped format that can be used by dateFns
  */
-const mapDateFormat = dateFormatString => {
+export const mapDateFormat = dateFormatString => {
     const dateMap = new Map(
         [
             ['%a', 'ddd'],
@@ -82,7 +83,7 @@ const mapDateFormat = dateFormatString => {
  * @param {Array} array - array with strings
  * @returns {Array} array with unique strings
  */
-const arrayUnique = array => {
+export const arrayUnique = array => {
     var a = array.concat();
     for (let i=0; i < a.length; ++i) {
         for (let j = i + 1; j < a.length; ++j) {
@@ -100,7 +101,7 @@ const arrayUnique = array => {
  * @param {Array} eclude - array of strings which we want to exclude in baseArray
  * @returns {Array} reduced array
  */
-const arrayExclude = (baseArray, exclude) => {
+export const arrayExclude = (baseArray, exclude) => {
     let newArray = [];
     for (let i=0; i < baseArray.length; i++) {
         if (!exclude.includes(baseArray[i])) {
@@ -110,11 +111,29 @@ const arrayExclude = (baseArray, exclude) => {
     return newArray;
 }
 
-export {
-    combineQualities,
-    humanFileSize,
-    mapDateFormat,
-    isDevelopment,
-    arrayUnique,
-    arrayExclude
+/**
+ * A simple wait function.
+ * @param {number} ms - Time to wait.
+ * @returns {Promise<number>} Resolves when done waiting.
+ */
+export const wait = /* istanbul ignore next */ ms => new Promise(resolve => setTimeout(resolve, ms));
+
+/**
+ * Returns when `check` evaluates as truthy.
+ * @param {function} check - Function to evaluate every poll interval.
+ * @param {number} [poll=100] - Interval to check, in milliseconds.
+ * @param {number} [timeout=3000] - Timeout to stop waiting after, in milliseconds.
+ * @returns {Promise<number>} The approximate amount of time waited, in milliseconds.
+ * @throws Will throw an error when the timeout has been exceeded.
+ */
+export const waitFor = /* istanbul ignore next */ async (check, poll = 100, timeout = 3000) => {
+    let ms = 0;
+    while (!check()) {
+        await wait(poll); // eslint-disable-line no-await-in-loop
+        ms += poll;
+        if (ms > timeout) {
+            throw new Error(`waitFor timed out (${timeout}ms)`);
+        }
+    }
+    return ms;
 };